Upgrade to Hadoop 0.9.1 .
git-svn-id: https://svn.apache.org/repos/asf/lucene/nutch/trunk@483420 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/nutch/crawl/CrawlDb.java
@@ -115,11 +115,13 @@ public class CrawlDb extends ToolBase {
     FileSystem fs = new JobClient(job).getFs();
     Path old = new Path(crawlDb, "old");
     Path current = new Path(crawlDb, CrawlDatum.DB_DIR_NAME);
-    fs.delete(old);
-    fs.rename(current, old);
+    if (fs.exists(current)) {
+      if (fs.exists(old)) fs.delete(old);
+      fs.rename(current, old);
+    }
     fs.mkdirs(crawlDb);
     fs.rename(newCrawlDb, current);
-    fs.delete(old);
+    if (fs.exists(old)) fs.delete(old);
   }
 
   public static void main(String[] args) throws Exception {
